# docs/swagger.yaml
openapi: 3.0.3
info:
  title: BalStorage API
  version: 1.0.0
  description: API documentation for BalStorage cloud storage.
servers:
  - url: https://storage.iqbalpradipta.my.id/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /folders/{id}/files:
    get:
      tags: [Files]
      summary: List files in folder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Files fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFilesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Files]
      summary: Upload one or more files
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Files uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /files/{id}:
    get:
      tags: [Files]
      summary: Get file detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: File fetched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/File"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Files]
      summary: Rename file
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameFileRequest"
      responses:
        "200":
          description: File renamed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/File"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Files]
      summary: Soft delete file
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
          format: int64
...
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        folder_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        original_name:
          type: string
        stored_name:
          type: string
        mime_type:
          type: string
        size:
          type: integer
          format: int64
        discord_message_id:
          type: string
        discord_attachment_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    RenameFileRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: new-name.png
...
    PaginatedFilesResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: "#/components/schemas/File"
                pagination:
                  $ref: "#/components/schemas/Pagination"
...
    UploadResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                data:
                  type: array
                  items:
                    type: object
                    properties:
                      original_name:
                        type: string
                      success:
                        type: boolean
                      error:
                        type: string
                      file:
                        $ref: "#/components/schemas/File"
                total:
                  type: integer
                success:
                  type: integer
                failed:
                  type: integer